+++ padaria-app/src/app/module/pages/manufacturers/src/manufacturers.component.ts
@@ -5,9 +5,9 @@ import { ManufacturerService } from '@module/services';
 import { ModalComponent } from '@module/shared/src';
 import { FormGridCommandEventArgs } from '@module/shared/src/form-grid/formgrid.component';
 import { SfGridColumnModel, SfGridColumns } from '@module/shared/src/grid';
-import { untilDestroyed } from '@module/utils/common';
+import { untilDestroyed, untilDestroyedAsync } from '@module/utils/common';
 import { markAllAsTouched } from '@module/utils/forms';
-import { ToastService } from '@module/utils/services';
+import { MessageService, ToastService } from '@module/utils/services';
 import { FormValidators } from '@syncfusion/ej2-angular-inputs';
 
 const NEW_ID = 'NOVO';
@@ -22,7 +22,6 @@ interface GridRow {
   templateUrl: './manufacturers.component.html',
 })
 export class ManufacturersComponent implements OnInit, OnDestroy {
-
   @ViewChild(ModalComponent, { static: true })
   modal!: ModalComponent;
 
@@ -32,6 +31,7 @@ export class ManufacturersComponent implements OnInit, OnDestroy {
 
   constructor(
     private toastService: ToastService,
+    private messageService: MessageService,
     private manufacturerService: ManufacturerService
   ) {}
 
@@ -50,6 +50,16 @@ export class ManufacturersComponent implements OnInit, OnDestroy {
     }
     const model = this.getModel();
     const exists = model.id > 1;
+
+    if (exists) {
+      const confirmed$ = this.messageService.showConfirmSave();
+      const confirmed = await untilDestroyedAsync(
+        confirmed$.asObservable(),
+        this
+      );
+      if (!confirmed) return;
+    }
+
     if (
       (exists
         ? this.manufacturerService.updateById(model)
@@ -106,12 +116,19 @@ export class ManufacturersComponent implements OnInit, OnDestroy {
   }
 
   private async onCommandRemove(model: GridRow): Promise<void> {
+    const confirmed$ = this.messageService.showConfirmSave();
+    const confirmed = await untilDestroyedAsync(
+      confirmed$.asObservable(),
+      this
+    );
+    if (!confirmed) return;
+
     this.manufacturerService
       .deleteById(model.id)
       .pipe(untilDestroyed(this))
       .subscribe(
         () => {
-          this.toastService.showSuccess('Excluído com sucesso!');
+          this.toastService.showRemove();
           this.loadData();
         },
         (error) => this.toastService.showError()
@@ -142,7 +159,7 @@ export class ManufacturersComponent implements OnInit, OnDestroy {
   private populateForm(manufacturer: Manufacturer): void {
     this.form.patchValue({
       id: manufacturer.id,
-      name: manufacturer.name
+      name: manufacturer.name,
     });
   }
 
@@ -165,7 +182,7 @@ export class ManufacturersComponent implements OnInit, OnDestroy {
       id: new FormControl({ value: NEW_ID, disabled: true }),
       name: new FormControl(null, [
         FormValidators.required,
-        Validators.maxLength(200)
+        Validators.maxLength(200),
       ]),
     }));
   }
@@ -173,7 +190,7 @@ export class ManufacturersComponent implements OnInit, OnDestroy {
   private createColumns(): SfGridColumnModel[] {
     return SfGridColumns.build<GridRow>({
       id: SfGridColumns.text('id', 'Código').minWidth(100).isPrimaryKey(true),
-      name: SfGridColumns.text('name', 'Nome').minWidth(200)
+      name: SfGridColumns.text('name', 'Nome').minWidth(200),
     });
   }
 }
