+++ padaria-app/src/app/module/pages/products/src/unit-measure/unit-measure.component.ts
@@ -5,9 +5,9 @@ import { UnitMeasureService } from '@module/services';
 import { ModalComponent } from '@module/shared/src';
 import { FormGridCommandEventArgs } from '@module/shared/src/form-grid/formgrid.component';
 import { SfGridColumnModel, SfGridColumns } from '@module/shared/src/grid';
-import { untilDestroyed } from '@module/utils/common';
+import { untilDestroyed, untilDestroyedAsync } from '@module/utils/common';
 import { markAllAsTouched } from '@module/utils/forms';
-import { ToastService } from '@module/utils/services';
+import { MessageService, ToastService } from '@module/utils/services';
 import { FormValidators } from '@syncfusion/ej2-angular-inputs';
 
 const NEW_ID = 'NOVO';
@@ -32,7 +32,8 @@ export class UnitMeasureComponent implements OnInit, OnDestroy {
 
   constructor(
     private toastService: ToastService,
-    private unitMeasureService: UnitMeasureService
+    private unitMeasureService: UnitMeasureService,
+    private messageService: MessageService
   ) {}
 
   ngOnInit(): void {
@@ -60,6 +61,16 @@ export class UnitMeasureComponent implements OnInit, OnDestroy {
     }
     const model = this.getModel();
     const exists = model.id > 1;
+
+    if (exists) {
+      const confirmed$ = this.messageService.showConfirmSave();
+      const confirmed = await untilDestroyedAsync(
+        confirmed$.asObservable(),
+        this
+      );
+      if (!confirmed) return;
+    }
+
     if (
       (exists
         ? this.unitMeasureService.updateById(model)
@@ -106,6 +117,13 @@ export class UnitMeasureComponent implements OnInit, OnDestroy {
   }
 
   private async onCommandRemove(model: GridRow): Promise<void> {
+    const confirmed$ = this.messageService.showConfirmSave();
+    const confirmed = await untilDestroyedAsync(
+      confirmed$.asObservable(),
+      this
+    );
+    if (!confirmed) return;
+
     this.unitMeasureService
       .deleteById(model.id)
       .pipe(untilDestroyed(this))
@@ -143,7 +161,7 @@ export class UnitMeasureComponent implements OnInit, OnDestroy {
     this.form.patchValue({
       id: unitMeasure.id,
       name: unitMeasure.name,
-      abbreviation: unitMeasure.abbreviation
+      abbreviation: unitMeasure.abbreviation,
     });
   }
 
@@ -167,9 +185,9 @@ export class UnitMeasureComponent implements OnInit, OnDestroy {
       id: new FormControl({ value: NEW_ID, disabled: true }),
       name: new FormControl(null, [
         FormValidators.required,
-        Validators.maxLength(200)
+        Validators.maxLength(200),
       ]),
-      abbreviation: new FormControl(null, Validators.maxLength(5))
+      abbreviation: new FormControl(null, Validators.maxLength(5)),
     }));
   }
 
@@ -177,7 +195,7 @@ export class UnitMeasureComponent implements OnInit, OnDestroy {
     return SfGridColumns.build<GridRow>({
       id: SfGridColumns.text('id', 'Código').minWidth(100).isPrimaryKey(true),
       abbreviation: SfGridColumns.text('abbreviation', 'Sigla').minWidth(200),
-      name: SfGridColumns.text('name', 'Nome').minWidth(200)
+      name: SfGridColumns.text('name', 'Nome').minWidth(200),
     });
   }
 }
